api/requirements.txt
asgiref==3.4.1
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
importlib-metadata==4.8.3
Markdown==3.3.6
pkg-resources==0.0.0
PyJWT==1.7.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
zipp==3.6.0
